feat: abandoned n8n, implemented MVP using typescript and nodejs.
n8n makes it really hard to build if you know how to code. ultimately I did develop an n8n based mvp but specifically the inability to pass variables to workflows via process.env or similar was a dealbreaker.  also the inability for credentials to be easily passed in.

this new prototype works end to end, so far only a single discord server is supported, pulling messages from a list of channel ID's passed in via env var.

using a fast and cheap LLM via API for now and posting to a single target (slack initially).

stay tuned!